+++ gastropaysdk/src/main/java/com/inventiv/gastropaysdk/data/response/NotificationPreferencesResponse.kt
@@ -1,0 +1,11 @@
+package com.inventiv.gastropaysdk.data.response
+
+import com.inventiv.gastropaysdk.ui.settings.notification.NotificationPreferencesChannelType
+import com.inventiv.gastropaysdk.ui.settings.notification.NotificationPreferencesStateType
+
+internal data class NotificationPreferencesResponse(
+    var id: Int,
+    var label: String,
+    var channel: NotificationPreferencesChannelType,
+    var state: NotificationPreferencesStateType
+)