+++ gastropaysdk/src/main/java/com/inventiv/gastropaysdk/repository/WalletRepositoryImp.kt
@@ -40,4 +40,6 @@ internal class WalletRepositoryImp(private val gastroPayService: GastroPayServic
             gastroPayService.invoiceSend(request)
         }
     }
+
+
 }