+++ gastropaysdk/src/main/java/com/inventiv/gastropaysdk/ui/settings/SettingsFragment.kt
@@ -15,8 +15,8 @@ import com.inventiv.gastropaysdk.shared.GastroPaySdk
 import com.inventiv.gastropaysdk.ui.MainViewModel
 import com.inventiv.gastropaysdk.ui.MainViewModelFactory
 import com.inventiv.gastropaysdk.ui.contactus.ContactUsFragment
+import com.inventiv.gastropaysdk.ui.settings.notification.NotificationPreferencesFragment
 import com.inventiv.gastropaysdk.ui.webview.WebViewFragment
-import com.inventiv.gastropaysdk.utils.blankj.utilcode.util.LogUtils
 import com.inventiv.gastropaysdk.utils.blankj.utilcode.util.StringUtils
 import com.inventiv.gastropaysdk.utils.delegate.viewBinding
 import com.inventiv.gastropaysdk.utils.getSettings
@@ -62,10 +62,14 @@ internal class SettingsFragment : BaseFragment(R.layout.fragment_settings_gastro
         binding.termsButton.setOnClickListener {
             viewModel.getTerms()
         }
+        binding.permissionsButton.setOnClickListener {
+            sharedViewModel.pushFragment(NotificationPreferencesFragment())
+
+        }
         binding.contactUsButton.setOnClickListener {
             sharedViewModel.pushFragment(ContactUsFragment())
         }
-        binding.faqButton.setOnClickListener{
+        binding.faqButton.setOnClickListener {
             getSettings().faq?.let { url ->
                 sharedViewModel.pushFragment(
                     WebViewFragment.newInstance(
@@ -112,7 +116,6 @@ internal class SettingsFragment : BaseFragment(R.layout.fragment_settings_gastro
                 when (it) {
                     is MainViewModel.ColdEvent.OnGenericWebViewClick -> {
                         sharedViewModel.resetColdEvent()
-                        LogUtils.d("GENERIC_WEBVIEW_IS_ACCEPT", it.isAccept)
                     }
                     else -> {
                     }