+++ gastropaysdk/src/main/java/com/inventiv/gastropaysdk/ui/settings/notification/NotificationPreferencesBase.kt
@@ -1,0 +1,20 @@
+package com.inventiv.gastropaysdk.ui.settings.notification
+
+import android.os.Parcelable
+import kotlinx.parcelize.Parcelize
+
+sealed class NotificationPreferencesBase {
+    @Parcelize
+    data class NotificationPreferencesHeader(
+        val id: Int,
+        var label: String
+    ) : NotificationPreferencesBase(), Parcelable
+
+    @Parcelize
+    data class NotificationPreferencesItem(
+        val id: Int,
+        var preferencesChannel: NotificationPreferencesChannelType,
+        var preferencesState: NotificationPreferencesStateType
+    ) : NotificationPreferencesBase(), Parcelable
+}
+