+++ gastropaysdk/CHANGELOG.md
@@ -9,6 +9,7 @@ Tags: [Added, Changed, Deprecated, Removed, Fixed, Security]
 ## [Unreleased](https://github.com/multinetinventiv/GastroPay-SDK-Android/compare/)
 
 ### Added
+- Notification Preferences [#2033](https://dev.azure.com/inventivtech/Panda/_workitems/edit/2033)
 - Settings Faq [#2031](https://dev.azure.com/inventivtech/Panda/_workitems/edit/2031)
 - Settings Contact Us [#2034](https://dev.azure.com/inventivtech/Panda/_workitems/edit/2034)
 - Wallet Detail [#130423](http://isttfs02:8080/tfs/MultinetCollection/Prj%20-%20%C4%B0sfanbul/_workitems/edit/130423)